% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_latex.R
\name{escape_latex}
\alias{escape_latex}
\title{Escape LaTeX Strings}
\usage{
escape_latex(x, newlines = FALSE, spaces = FALSE)
}
\arguments{
\item{x}{a string to be escaped}

\item{newlines}{if \code{TRUE} new line characters are escaped as a double
backslash}

\item{spaces}{if \code{TRUE} multiple spaces are escaped to multiple
\emph{protected} whitespaces in LaTeX}
}
\value{
an string with LaTeX control characters escaped
}
\description{
The functions escapes characters that are used as LaTeX control sequences.
}
\details{
If \code{x} is of class \code{TeXcontent}, the input string is
 returned ``as is''.
}
\references{
The function was taken from

Yihui Xie (2016). knitr: A General-Purpose Package for Dynamic Report
Generation in R. R
}

