% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print.pgfkeyvalue}
\alias{knit_print.pgfkeyvalue}
\alias{knit_print.pgfvalue}
\title{Printing functions for \code{knitr}}
\usage{
\method{knit_print}{pgfkeyvalue}(x, options = list())

\method{knit_print}{pgfvalue}(x, options = list())
}
\arguments{
\item{x}{an object of S3 class \code{\link{pgfvalue}} or
\code{\link{pgfkeyvalue}}}

\item{options}{a string, a common path for all keys, change_dir if \code{TRUE} the path is set using the \code{/.cd}
handler of the LaTeX command}
}
\description{
The \code{knit_print} functions fot \code{\link{pgfvalue}} and
\code{\link{pgfkeyvalue}} return LaTeX-code to set key-value pairs using
LaTeX's \code{pgfkeys}-package.
}
\details{
For objects \code{x} of S3 class \code{pgfkeyvalue}
 \code{\link[knitr]{knit_print}} returns LaTeX code to set pgf key-value
 pairs using \code{\\pgfkeys{key = value}}.
 
 The optional argument \code{option} contain knitr chunk options. Available
 options for printing key value pairs are
 \describe{
   \item{\code{path}}{a common path to prepend pgf-keys}
   \item{\code{change_dir}}{logical, for \code{pgfkeyvalue} objects only;
     If \code{TRUE} the path is set globally by using \code{/.cd} handler}
   \item{\code{rm_unnamed}}{logical, for \code{pgfvalue} objects only;
     If \code{TRUE} unnamed arguments are omitted silently. If \code{FALSE}
     an error is thrown on any unnamed argument (default).}
 }

For objects \code{x} of S3 class \code{pgfkeyvalue}
 \code{\link[knitr]{knit_print}} returns LaTeX code to set pgf key-value
 pairs using \code{\\pgfkeyssetvalue{key}{value}}.
}

