% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgf_print.R
\name{pgf_print}
\alias{.pgfPath}
\alias{pgf_print}
\alias{pgf_print.data.frame}
\alias{pgf_print.default}
\alias{pgf_print.pgfkeyvalue}
\alias{pgf_print.pgfvalue}
\title{Printing functions for \code{knitr}}
\usage{
pgf_print(x, options, ...)

\method{pgf_print}{default}(x, options, ...)

\method{pgf_print}{pgfkeyvalue}(x, options, ...)

\method{pgf_print}{pgfvalue}(x, options, ...)

\method{pgf_print}{data.frame}(x, options, ...)

.pgfPath(options)
}
\arguments{
\item{x}{key value pairs to pass to pgf-keys}

\item{options}{list of current chunk options. For used chunk options, see
the "Chunk Options" section.}
}
\value{
The internal function \code{.pgfPath} returns the full path for
 pgfkeys with one attribute \code{change_dir}.
}
\description{
The \code{knit_print} functions fot \code{\link{pgfvalue}} and
\code{\link{pgfkeyvalue}} return LaTeX-code to set key-value pairs using
LaTeX's \code{pgfkeys}-package.

\code{.pgfPath} is an auxilliary function to get the pgf-keys'
 path.
}
\details{
Chunk Options:
\describe{
 \item{\code{setValue}}{if \code{TRUE} \code{\\pgfkeyssetvalue} is used
   instead of \code{\\pgfkeys}}
 \item{\code{path}}{path for pgf-keys}
 \item{\code{change_dir}}{if \code{TRUE} \code{.cd}-handler is used instead
   of setting path for each key explicitely}
 \item{\code{rm_unnamed}}{if \code{TRUE} unname values are omitted, if
   \code{FALSE} and names are missing, an error is returned}
 \item{\code{colSep}}{delimiter to separate columns in LaTeX tables}
 \item{\code{rowSep}}{delimiter to separater rows in LaTeX tables}
 \item{\code{tabName}}{name to store read table}
}

For objects \code{x} of S3 class \code{pgfkeyvalue}
 \code{\link[knitr]{knit_print}} returns LaTeX code to set pgf key-value
 pairs using \code{\\pgfkeyssetvalue{key}{value}}.
}
\keyword{internal}

