% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfKeys-package.R
\docType{package}
\name{pgfKeys-package}
\alias{pgfKeys-package}
\title{pgfKeys: A package for Setting pgf-Keys in LaTeX}
\description{
The package provide functions to format and output variables for the use
with pgf-keys in LaTeX
}

