% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfvalue.R
\docType{methods}
\name{pgfvalue}
\alias{pgfvalue}
\alias{pgfvalue,character-method}
\alias{pgfvalue,numeric-method}
\alias{pgfvalue,logical-method}
\alias{pgfvalue,POSIXt-method}
\alias{pgfvalue,Date-method}
\alias{pgfvalue,missing-method}
\alias{pgfvalue,list-method}
\alias{pgfvalue,pgfvalue-method}
\alias{pgfvalue,pgfkeyvalue-method}
\alias{pgfkeyvalue}
\title{PGF Key-Value Pairs}
\usage{
pgfvalue(x, ...)

\S4method{pgfvalue}{character}(x, ...)

\S4method{pgfvalue}{numeric}(x, ...)

\S4method{pgfvalue}{logical}(x, ...)

\S4method{pgfvalue}{POSIXt}(x, ...)

\S4method{pgfvalue}{Date}(x, ...)

\S4method{pgfvalue}{missing}(x, ...)

\S4method{pgfvalue}{list}(x, ...)

\S4method{pgfvalue}{pgfvalue}(x, ...)

\S4method{pgfvalue}{pgfkeyvalue}(x, ...)

pgfkeyvalue(...)
}
\arguments{
\item{x}{an object that can be coerced to a key-value pair; See signatures
on methods for available object classes.}

\item{...}{additional parameters, these can be named or unnamed, see details}
}
\value{
The methods \code{pgfvalue()} return an object of S3 class
 \code{pgfvalue}. It contains a character string of key-value pairs. If a
 name is given, it is the key's name and the entry is the corresponding
 value. If no name is given, the entry contains the key itself.

The function \code{pgfkeyvalue()} returns an object of S3 class
 \code{pgfkeyvalue}. The structure of this class is the same as for class
 \code{pgfvalue} except for a different way of printing output (see
 \code{\link[=knit_print.pgfvalue]{knit_print}} for details).
}
\description{
The methods generate key-value pairs as character strings.
}
\details{
All arguments (\code{x} as well as additional arguments \code{...})
 can be either named, i.e. \code{key = value} or unnamed, i.e. \code{key}.
 If an unnamed argument is a named vector (or list) itself it will be
 expanded to named arguments.

For \code{x} a character any special characters will be escaped,
 using \code{\link{escape_latex}}. If \code{x} contains special charcaters
 that should remain unescaped (e.g. if value contains LaTeX commands), the
 class of \code{x} should be set to \code{"pgfvalue"} explicitely before
 calling \code{pgfvalue()}, see Examples.

Logical values are given as \code{true} or \code{false}.

Lists can be nested. An unnamed list element is expanded to its
 (named or unnamed) entries. For a named element, the value is pasted to a
 single character with its values separated by a comma. The values (if named)
 are prepended by the values name and an equal sign. If the value contains
 a comma or an equal sign, it is masked with curly brackets.
}
\examples{
## unnamed arguments
pgfvalue('key 1', 'key 2')

## named arguments
pgfvalue(key_1 = 'value 1', key_2 = 'value 2')

## named vector
pgfvalue(c(key_1 = 'value 1', key_2 = 'value 2'))

## escaping special character
x <- c(key_1 = 'value_1', key_2 = '\\\\command{argument}')
pgfvalue(x)

class(x) <- 'pgfvalue'
pgfvalue(x)

## numeric values
pgfvalue(num_1 = 1.3, num_2 = 12.23)

## logical values
pgfvalue(bool = TRUE)

## POSIXt values
pgfvalue(start = as.POSIXct('2016-04-01'), end = Sys.time())

## simple list as unnamed argument
pgfvalue(list(num_key = 12.3, log_key = TRUE, chr_key = 'text value'))

## simple lists as named arguments
pgfvalue(list_1 = list(num = 2, val = 'text'),
 list_2 = list(date = Sys.Date(), description = 'current date'))

## nested lists
pgfvalue(list_1 = list(num = 2, val = 'text',
 list_2 = list(date = Sys.Date(), description = 'current date')))

}
