% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfvalue_print.R
\name{pgfvalue_print}
\alias{pgfvalue_print}
\alias{pgfvalue_print.list}
\alias{pgfvalue_print.InsuredOrPrincipal}
\title{Print PGF Key-Value Pairs}
\usage{
pgfvalue_print(x, options, ...)

\method{pgfvalue_print}{list}(x, options, ...)

\method{pgfvalue_print}{InsuredOrPrincipal}(x, options, ...)
}
\arguments{
\item{x}{object}

\item{options}{knitr options}

\item{...}{additional arguments}
}
\description{
This printing function can be used for dynamic LaTeX documents to set
PGF keys using \command{pgfkeyssetvalue}.
}
