% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_pgf.R
\name{render_pgf}
\alias{render_pgf}
\title{Set Output Hooks for knitr}
\usage{
render_pgf()
}
\description{
The function \code{render_pgf} can be used to set up output hooks for the
use with knitr. It is based on the \code{\link[knitr]{render_latex}()}
function of the knitr package and should therefore only be used with LaTeX
output format.
}
\section{Default Chunk Options}{

 The function \code{render_pgf} sets the chunk options \code{error = FALSE}
 and \code{message = FALSE} to prevent document generation with erroneous
 code.
}

\section{Output Hook Function}{

 The hook function \code{output} is extended by an additional boolean option
 \code{skip_lines}. If \code{TRUE} double line skips are removed and the
 line skip at the end of the output is removed to. This is used for the
 output of tables (which must not have an empty line at the end).
}

\section{Option Templates}{

 An option template \code{pgfkey} is defined. It can be used to output
 return of \code{\link{pgfvalue}} or \code{\link{pgfkeyvalue}}.
}

\section{Hook for Table Output}{

 The output hook-function \code{pgfplotstable} can be used to set input for
 tables using the LaTeX-command \code{\\pgfplotstableread}. The argument to
 the hook is the name of the LaTeX command the table will be stored in.
}

\section{Sample Code Chunk}{

 The following code chunk can be used at the beginning of the document to
 initialize pgf-rendering.
 
 \preformatted{<<setup, include = FALSE>>=
 render_pgf()
 @}
}
\seealso{
\code{\link{pgfvalue}}, \code{\link{pgfkeyvalue}} for a definition
 of the respective S3 classes.
 
 \code{\link[=knit_print.pgfvalue]{knit_print}} for printing method for
 \code{\link{pgfvalue}} and \code{\link{pgfkeyvalue}} class objects.
}

